+++ sql/materialized_views.sql
@@ -60,8 +60,6 @@ and tb2.parent=tb3.id;
 -- details about the school(both primary and preschools)
 -- putting the locations in a view to save query time
 -- assembly and parliament IDs as well.
-DROP MATERIALIZED VIEW IF EXISTS mvw_school_details;
-
 CREATE MATERIALIZED VIEW mvw_school_details as
 SELECT tbs.id as id,
     tb1.id as cluster_or_circle_id,
@@ -124,3 +122,17 @@ SELECT t1.dise_code,
     t1.score,
     t1.df_group
    FROM dblink('host=localhost dbname=dise_all user=klp password=klp'::text, 'select * from tb_dise_facility_agg'::text) t1(dise_code character varying(32), df_metric character varying(30), score numeric(5,0), df_group character varying(30));
+
+
+CREATE MATERIALIZED VIEW mvw_school_class_total_year AS
+SELECT sg.sid AS schid,
+    btrim(sg.name::text) AS clas,
+    count(DISTINCT stu.id) AS total,
+    acyear.id AS academic_year
+   FROM tb_student_class stusg,
+    tb_class sg,
+    tb_student stu,
+    tb_academic_year acyear
+  WHERE stu.id = stusg.stuid AND stusg.clid = sg.id AND stu.status = 2 AND acyear.id = stusg.ayid AND (acyear.name::text IN ( SELECT DISTINCT vw_lib_level_agg.year
+           FROM vw_lib_level_agg))
+  GROUP BY sg.sid, btrim(sg.name::text), acyear.id;